requirements.txt
requests>=2.31.0,<3.0.0
dnspython>=2.6.0,<3.0.0
beautifulsoup4>=4.12.0,<5.0.0
reportlab>=4.0.0,<5.0.0
shodan>=1.30.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
cryptography>=42.0.0,<43.0.0